Add files via upload
Domain Description
Allstate is a personal insurance company in the US. Allstate is currently developing automated methods of predicting the cost, and hence severity, of claims. This project provides an algorithm that accurately predicts claims severity.

Description of Algorithm
The algorithm used for predicting the Allstate claims severity is XGBoost. XGBoost is an algorithm that has recently been dominating applied machine learning and Kaggle competitions for structured or tabular data. XGBoost is an implementation of gradient boosted decision trees designed for speed and performance. It belongs to a broader collection of tools under the umbrella of the Distributed Machine Learning Community or DMLC who are also the creators of the popular mxnet deep learning library. The XGBoost algorithm was chosen because of its computational speed and model performance. It makes best use of available resources to train the model. Along with this, sklearn was also used. Specifically the packages